# prueba_tecnica_2/src/main/java/com/modusoftware/pruebaTecnica2/business/dto/VentaDTOGetAll.java
package com.modusoftware.pruebaTecnica2.business.dto;

import java.time.LocalDateTime;

import com.modusoftware.pruebaTecnica2.persistent.model.LibroEntity;

import lombok.Data;

@Data
public class VentaDTOGetAll {

	private Integer id;
	private Integer libroId;
	private String libroName;
	private LocalDateTime fechaVenta;
	private Integer cantidad;
	
	
	
	}
	